Organizing tests between rust files
