+++ src/core.rs
@@ -2,7 +2,7 @@ use std::fmt;
 use std::error;
 
 use crate::exp::LisperExp;
-use crate::env::{LisperEnv, create_default_env};
+use crate::env::{LisperEnv};
 // An error type for the Lisp interperter
 #[derive(Debug)]
 pub enum LisperErr {
@@ -321,392 +321,9 @@ mod tests {
     }
 
     #[test]
-    fn create_default_env_add() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("+").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 52.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 + arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_sub() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("-").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 52.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 - arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-        
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_mul() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("*").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 52.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 * arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_div() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("/").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 52.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 / arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_mod() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("%").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 52.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 % arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_less_than() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("<").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 5.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Bool(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 < arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_more_than() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get(">").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 5.0;
-                let arg1_f64: f64 = 13.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Bool(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 > arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-        
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_equals() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("=").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 5.0;
-                let arg1_f64: f64 = 5.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Bool(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 == arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_less_or_equal() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("<=").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 6.0;
-                let arg1_f64: f64 = 5.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Bool(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 <= arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_more_or_equal() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get(">=").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = 3.0;
-                let arg1_f64: f64 = 5.0;
-        
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-                let arg1:LisperExp = LisperExp::Number(arg1_f64);
-        
-                if let LisperExp::Bool(res) = f(&LisperExp::List(vec![arg0, arg1])) {
-                    assert_eq!(res, arg0_f64 >= arg1_f64);
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_sin() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("sin").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = core::f64::consts::PI;
-
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0])) {
-                    assert_eq!(res, arg0_f64.sin());
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_cos() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("cos").ok_or_else(|| 
-            LisperErr::Reason("Error, function not found.".to_string())
-        )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = core::f64::consts::PI;
-
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0])) {
-                    assert_eq!(res, arg0_f64.cos());
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn create_default_env_tan() -> Result<(),  Box<dyn std::error::Error>> {
-        use super::*;
-        
-        let env:LisperEnv = create_default_env();
-
-        let func = env.data.get("tan").ok_or_else(|| 
-                LisperErr::Reason("Error, function not found.".to_string())
-            )?;
-
-        match func {
-            LisperExp::Func(f) => {
-                let arg0_f64: f64 = core::f64::consts::PI;
-
-                let arg0:LisperExp = LisperExp::Number(arg0_f64);
-        
-                if let LisperExp::Number(res) = f(&LisperExp::List(vec![arg0])) {
-                    assert_eq!(res, arg0_f64.tan());
-                } else {
-                    assert!(false);
-                }
-            },
-            _ => assert!(false)
-        }
-
-        Ok(())
-    }
-
-    #[test]
     fn eval_def() -> Result<(), Box<dyn std::error::Error>> {
         use super::*;
+        use crate::env::create_default_env;
 
         // Test if eval handles def
         // Format: (def variable_name[as LisperExp::Symbol] (value[as LisperExp]))
@@ -741,6 +358,7 @@ mod tests {
     #[test]
     fn eval_if() -> Result<(), Box<dyn std::error::Error>> {
         use super::*;
+        use crate::env::create_default_env;
 
         // Test if eval handles if
         // Format: (if (if expression[as LisperExp]) (true expression[as LisperExp]) (false expression[as LisperExp]))
@@ -771,6 +389,7 @@ mod tests {
     #[test]
     fn eval_fn() -> Result<(), Box<dyn std::error::Error>> {
         use super::*;
+        use crate::env::create_default_env;
 
         // Test if eval handles fn
         // Format: (fn function_name[as string] (argument[as LisperExp list]) (function[as LisperExp]))
